feat: move transaction history related to core
Closes #1178
